requirements/base.txt
Django==1.9
djangorestframework==3.4.6
djangorestframework-xml==1.3.0
Pillow==3.3.1
